+++ src/WebApp/Pages/Shared/_Layout.cshtml
@@ -68,5 +68,6 @@
       const dateTimeIntervalId = setInterval(updateDateTime, 1000);
       updateDateTime();
     </script>
+    @RenderSection("Scripts", required: false)
 </body>
 </html>
